Allow configuring a timeout for requests
1. Make the timout explicit for gunicorn workers in the Dockerfile.

2. Extend configuration with a 'REQUEST_TIMEOUT' option, which should be
considered both for Koji and Quay requests. For the later, the only
exception is the push operation, as Operator Courier API does not
have a timeout option yet.

All the above is done, so that the OMPS can be configured in a way,
where requests time out sooner than the workers. This way we hope to
catch some issues currently hidden by the worker timeouts.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>