Tests: adjust some assertions
With #97 the message when reaching out to a quay org not accessible
changed: adjust the test expectation accordingly.

When deleting all versions, the way versions are sorted in the deleted
list changed: make sure both the actual and expected lists are sorted to
avoid tests failing for this reason in the future.

Remove an unused environment variable from the tox config.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>