Remove rpm operator-courier dependency and install it from pypi instead
When there is a new version of courier, the OCP Group F team has to
release it to PyPI, and then we pull it down and build a fedora rawhide
rpm out of it, and then we build a fedora 30+ rpm out of it and create
a bodhi update for that. Then, we wait ~24 hours for it to land in the
testing repo, so that we can build our container image and start
testing OMPS with the new version. This is a silly delay.